Tratamento de exceções - delete

+++ src/main/java/com/larrieu/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.larrieu.course.entities.User;
 import com.larrieu.course.repositories.UserRepository;
+import com.larrieu.course.services.exceptions.DatabaseException;
 import com.larrieu.course.services.exceptions.ResourceNotFoundException;
 
 
@@ -29,8 +31,18 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
+	
+	
 	//diferentemente do findbyid, o getreference MONITORA o objeto
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
